requirements.txt
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
antlr4-python3-runtime==4.9.3
anyio==4.13.0
async-timeout==5.0.1
attrs==26.1.0
backports.zstd
Brotli
causal-conv1d==1.2.0.post2
certifi==2026.2.25
charset-normalizer==3.4.7
cuda-bindings==13.2.0
cuda-pathfinder==1.5.1
cuda-toolkit==13.0.2
datasets==4.8.4
dill==0.4.1
einops==0.8.2
exceptiongroup==1.3.1
filelock==3.25.2
frozenlist==1.8.0
fsspec==2026.3.0
gmpy2
h11==0.16.0
h2
hf-xet==1.4.3
hpack
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
hydra-core==1.3.2
hyperframe
idna==3.11
iopath==0.1.10
Jinja2==3.1.6
mamba-ssm==1.2.2
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
networkx==3.4.2
ninja==1.13.0
numpy==1.26.4
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
omegaconf==2.3.0
opencv-python-headless==4.11.0.86
packaging==26.0
pandas==2.3.3
peft==0.18.1
pillow==12.2.0
portalocker==3.2.0
propcache==0.4.1
psutil==7.2.2
pyarrow==23.0.1
PySocks
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
safetensors==0.7.0
sentencepiece==0.2.1
six==1.17.0
sympy==1.14.0
timm==1.0.26
tokenizers==0.21.4
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.67.3
transformers==4.49.0
triton==3.6.0
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
xxhash==3.6.0
yarl==1.23.0
